pyproject.toml
[tool.poetry]
name = "courses-drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
black = "^24.8.0"
isort = "^5.13.2"
ipython = "^8.26.0"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
